requirements.txt
asgiref==3.8.1
atomicwrites==1.4.1
attrs==24.2.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==2.0.12
colorama==0.4.6
cryptography==43.0.0
defusedxml==0.8.0rc2
Django==4.2.15
django-allauth==64.1.0
django-filter==2.4.0
django-rest-framework==0.1.0
django-restframework==0.0.1
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
flake8==7.1.1
flake8-docstrings==1.7.0
idna==3.8
iniconfig==2.0.0
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.1
Pillow==9.3.0
pluggy==0.13.1
py==1.11.0
pycodestyle==2.12.1
pycparser==2.22
pydocstyle==6.3.0
pyflakes==3.2.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python3-openid==3.2.0
pytz==2024.1
requests==2.26.0
requests-oauthlib==2.0.0
snowballstemmer==2.2.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
toml==0.10.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
